content/html/content/test/test_bug559284.html


<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=559284
-->
<head>
  <title>Test for Bug 559284</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=559284">Mozilla Bug 559284</a>
<p id="display"></p>
<div id="content" style="display: none">
  <section id="section"></section>
  <nav id="nav"></nav>
  <article id="article"></article>
  <aside id="aside"></aside>
  <hgroup id="hgroup"></hgroup>
  <header id="header"></header>
  <footer id="footer"></footer>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 559284 **/
  todo_is($("section").constructor, "[object HTMLElement]",
          "Constructor should be section");
  todo_is($("section").toString().substring(0, 19), "[object HTMLElement",
          "Section should be HTMLElement");
  is($("section") instanceof HTMLElement, true,
       "Section should be instance of HTMLElement");
  todo_is($("nav").constructor, "[object HTMLElement]",
          "Constructor should be nav");
  todo_is($("nav").toString().substring(0, 19), "[object HTMLElement",
          "Nav should be HTMLElement");
  is($("nav") instanceof HTMLElement, true,
       "Nav should be instance of HTMLElement");
  todo_is($("article").constructor, "[object HTMLElement]",
          "Constructor should be article");
  todo_is($("article").toString().substring(0, 19), "[object HTMLElement",
          "Article should be HTMLElement");
  is($("article") instanceof HTMLElement, true,
       "Article should be instance of HTMLElement");
  todo_is($("aside").constructor, "[object HTMLElement]",
          "Constructor should be aside");
  todo_is($("aside").toString().substring(0, 19), "[object HTMLElement",
          "Aside should be HTMLElement");
  is($("aside") instanceof HTMLElement, true,
       "Aside should be instance of HTMLElement");
  todo_is($("hgroup").constructor, "[object HTMLElement]",
          "Constructor should be hgroup");
  todo_is($("hgroup").toString().substring(0, 19), "[object HTMLElement",
          "Hgroup should be HTMLElement");
  is($("hgroup") instanceof HTMLElement, true,
       "Hgroup should be instance of HTMLElement");
  todo_is($("header").constructor, "[object HTMLElement]",
          "Constructor should be header");
  todo_is($("header").toString().substring(0, 19), "[object HTMLElement",
          "Header should be HTMLElement");
  is($("header") instanceof HTMLElement, true,
       "Header should be instance of HTMLElement");
  todo_is($("footer").constructor, "[object HTMLElement]",
          "Constructor should be footer");
  todo_is($("footer").toString().substring(0, 19), "[object HTMLElement",
          "Footer should be HTMLElement");
  is($("footer") instanceof HTMLElement, true,
       "Footer should be instance of HTMLElement");
</script>
</pre>
</body>
</html>
